# Some changes to Mistral querier to get better CPU performance