--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -77,6 +77,13 @@ in development
   support non-ascii (unicode) characters. (bug fix)
 * When RBAC is enabled and action is scheduled (ran) through the API, include ``rbac`` dictionary
   with ``user`` and ``roles`` ``action_context`` attribute. (improvement)
+* Fix a bug in query base module when outstanding queries to mistral or other workflow engines 
+  could cause a tight loop without cooperative yield leading to 100% CPU usage by st2resultstracker
+  process. (bug-fix)
+* Make the query interval to third party workflow systems (including mistral) a configurable
+  value. You can now set ``query_interval`` in ``[results_tracker]`` section in ``/etc/st2/st2.conf``.
+  With this, the default query interval is set to 20s as opposed to 0.1s which was rather aggressive
+  and could cause CPU churn when there is a large number of outstanding workflows. (improvement)
 
 2.2.1 - April 3, 2017
 ---------------------