--- a/conf/st2.dev.conf
+++ b/conf/st2.dev.conf
@@ -100,3 +100,9 @@ logging = st2exporter/conf/logging.exporter.conf
 
 [content]
 pack_group = stanley
+
+[results_tracker]
+query_interval = 0.1
+
+[mistral]
+jitter_interval = 0