--- a/conf/st2.tests.conf
+++ b/conf/st2.tests.conf
@@ -86,3 +86,6 @@ logging = st2actions/conf/logging.notifier.conf
 
 [exporter]
 logging = st2exporter/conf/logging.exporter.conf
+
+[results_tracker]
+query_interval = 0.1