--- a/conf/st2.tests.conf
+++ b/conf/st2.tests.conf
@@ -89,3 +89,6 @@ logging = st2exporter/conf/logging.exporter.conf
 
 [results_tracker]
 query_interval = 0.1
+
+[mistral]
+jitter_interval = 0