--- a/st2tests/st2tests/config.py
+++ b/st2tests/st2tests/config.py
@@ -79,6 +79,7 @@ def _override_common_opts():
     CONF.set_override(name='mask_secrets', override=True, group='log')
     CONF.set_override(name='url', override='zake://', group='coordination')
     CONF.set_override(name='lock_timeout', override=1, group='coordination')
+    CONF.set_override(name='jitter_interval', override=0, group='mistral')
 
 
 def _override_api_opts():