@Kami 2017-05-17
Thanks and good catch both (@enykeev and @lakshmi-kannan) :+1: 

I'm personally with @enykeev and I think it would be good to randomly spread out requests a bit (to avoid thundering herd affect and overloading Mistral API):

> @lakstorm can we actually spread the requests around this 20s window instead of bursting them all at once?