@Kami 2017-05-22
Please update changelog.

Besides that, LGTM :+1: