@armab 2017-05-22
Because it's a forked PR, some integration checks like st2-packages + e2e were not triggered :/

Just a reminder to use a branch instead when possible.