@lakshmi-kannan 2017-05-22
Updated CHANGELOG. Also, docs are here https://github.com/StackStorm/st2docs/pull/503